refactor: organized modules into proper dir structure
