+++ backend/src/models/types/index.ts
@@ -1,0 +1,2 @@
+export type AutoSetFields = "createdOn" | "lastModifiedOn" | "ttl";
+export type ApiResponseUrlType = "GS_PATH" | "API_URI";
