+++ backend/src/pages/api/profile/[uid]/read/index.ts
@@ -1,10 +1,10 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import Identity, { PsudoFields, SchemaFields } from "@/models/Identity";
-import { respond } from "@/lib/utils/respond";
-import { withmiddleware } from "@/middlewares/withMiddleware";
-import { getLoggedInUser } from "@/middlewares/auth";
-import { CustomApiError } from "@/lib/utils/ApiError";
-import { RateLimits } from "@/middlewares/rateLimit";
+import { respond } from "@/utils/respond";
+import { WithMiddleware } from "@/middlewares/WithMiddleware";
+import { getLoggedInUser } from "@/middlewares/Auth";
+import { CustomApiError } from "@/types/CustomApiError";
+import { RateLimits } from "@/middlewares/RateLimiter";
 
 /**
  * ```
@@ -44,7 +44,7 @@ import { RateLimits } from "@/middlewares/rateLimit";
  * }
  * ```
  */
-export default withmiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
+export default WithMiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
   // Allow only GET requests
   if (req.method !== "GET") {
     throw CustomApiError.create(405, "Method Not Allowed");
