+++ backend/src/pages/api/profile/[uid]/updateType/index.ts
@@ -1,10 +1,10 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { respond } from "@/lib/utils/respond";
-import { authenticate } from "@/middlewares/auth";
+import { respond } from "@/utils/respond";
+import { authenticate } from "@/middlewares/Auth";
 import Identity, { IdentityType } from "@/models/Identity";
-import { withmiddleware } from "@/middlewares/withMiddleware";
-import { CustomApiError } from "@/lib/utils/ApiError";
-import { RateLimits } from "@/middlewares/rateLimit";
+import { WithMiddleware } from "@/middlewares/WithMiddleware";
+import { CustomApiError } from "@/types/CustomApiError";
+import { RateLimits } from "@/middlewares/RateLimiter";
 
 /**
  * ```
@@ -14,7 +14,7 @@ import { RateLimits } from "@/middlewares/rateLimit";
  * response = { message: string }
  * ```
  */
-export default withmiddleware(async function PATCH(req: NextApiRequest, res: NextApiResponse) {
+export default WithMiddleware(async function PATCH(req: NextApiRequest, res: NextApiResponse) {
   // Only allow PATCH method
   if (req.method !== "PATCH") {
     throw CustomApiError.create(405, "Method Not Allowed");
