+++ backend/src/pages/api/rooms/[roomId]/[imageIdOrUid]/updateRating/index.ts
@@ -1,11 +1,11 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { respond } from "@/lib/utils/respond";
-import { withmiddleware } from "@/middlewares/withMiddleware";
-import { CustomApiError } from "@/lib/utils/ApiError";
+import { respond } from "@/utils/respond";
+import { WithMiddleware } from "@/middlewares/WithMiddleware";
+import { CustomApiError } from "@/types/CustomApiError";
 import RoomRatings from "@/models/RoomRatings";
-import { getLoggedInUser } from "@/middlewares/auth";
+import { getLoggedInUser } from "@/middlewares/Auth";
 import Room, { SchemaFields } from "@/models/Room";
-import { RateLimits } from "@/middlewares/rateLimit";
+import { RateLimits } from "@/middlewares/RateLimiter";
 
 /**
  * ```
@@ -15,7 +15,7 @@ import { RateLimits } from "@/middlewares/rateLimit";
  * response = { message: string }
  * ```
  */
-export default withmiddleware(async function PATCH(req: NextApiRequest, res: NextApiResponse) {
+export default WithMiddleware(async function PATCH(req: NextApiRequest, res: NextApiResponse) {
   // Only allow PATCH method
   if (req.method !== "PATCH") {
     throw CustomApiError.create(405, "Method Not Allowed");
