+++ backend/src/pages/api/rooms/[roomId]/delete/index.ts
@@ -1,9 +1,9 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { respond } from "@/lib/utils/respond";
-import { getLoggedInUser } from "@/middlewares/auth";
-import { withmiddleware } from "@/middlewares/withMiddleware";
-import { CustomApiError } from "@/lib/utils/ApiError";
-import { RateLimits } from "@/middlewares/rateLimit";
+import { respond } from "@/utils/respond";
+import { getLoggedInUser } from "@/middlewares/Auth";
+import { WithMiddleware } from "@/middlewares/WithMiddleware";
+import { CustomApiError } from "@/types/CustomApiError";
+import { RateLimits } from "@/middlewares/RateLimiter";
 import Room, { SchemaFields } from "@/models/Room";
 
 /**
@@ -12,7 +12,7 @@ import Room, { SchemaFields } from "@/models/Room";
  * response = { message: string }
  * ```
  */
-export default withmiddleware(async function DELETE(req: NextApiRequest, res: NextApiResponse) {
+export default WithMiddleware(async function DELETE(req: NextApiRequest, res: NextApiResponse) {
   // Only allow DELETE method
   if (req.method !== "DELETE") {
     throw CustomApiError.create(405, "Method Not Allowed");
