+++ backend/src/pages/api/rooms/[roomId]/read/index.ts
@@ -1,10 +1,10 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import Room, { SchemaFields } from "@/models/Room";
-import { respond } from "@/lib/utils/respond";
-import { withmiddleware } from "@/middlewares/withMiddleware";
-import { getLoggedInUser } from "@/middlewares/auth";
-import { CustomApiError } from "@/lib/utils/ApiError";
-import { RateLimits } from "@/middlewares/rateLimit";
+import { respond } from "@/utils/respond";
+import { WithMiddleware } from "@/middlewares/WithMiddleware";
+import { getLoggedInUser } from "@/middlewares/Auth";
+import { CustomApiError } from "@/types/CustomApiError";
+import { RateLimits } from "@/middlewares/RateLimiter";
 
 /**
  * ```
@@ -32,7 +32,7 @@ import { RateLimits } from "@/middlewares/rateLimit";
  * }
  * ```
  */
-export default withmiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
+export default WithMiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
   if (!(await RateLimits.ROOM_READ(req, res))) return;
 
   // Allow only GET requests
