+++ backend/src/pages/api/rooms/[roomId]/readRating/index.ts
@@ -1,9 +1,9 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { respond } from "@/lib/utils/respond";
-import { withmiddleware } from "@/middlewares/withMiddleware";
-import { CustomApiError } from "@/lib/utils/ApiError";
+import { respond } from "@/utils/respond";
+import { WithMiddleware } from "@/middlewares/WithMiddleware";
+import { CustomApiError } from "@/types/CustomApiError";
 import Room, { SchemaFields } from "@/models/Room";
-import { RateLimits } from "@/middlewares/rateLimit";
+import { RateLimits } from "@/middlewares/RateLimiter";
 
 /**
  * ```
@@ -11,7 +11,7 @@ import { RateLimits } from "@/middlewares/rateLimit";
  * response = { rating: 0|1|2|3|4|5 }
  * ```
  */
-export default withmiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
+export default WithMiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
   if (!(await RateLimits.ROOM_RATING_READ(req, res))) return;
 
   // Only allow GET method
