+++ backend/src/pages/api/rooms/[roomId]/updateParams/index.ts
@@ -1,14 +1,14 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { respond } from "@/lib/utils/respond";
-import { getLoggedInUser } from "@/middlewares/auth";
+import { respond } from "@/utils/respond";
+import { getLoggedInUser } from "@/middlewares/Auth";
 import Identity, { SchemaFields as IdentitySchemaFields } from "@/models/Identity";
-import { withmiddleware } from "@/middlewares/withMiddleware";
-import { CustomApiError } from "@/lib/utils/ApiError";
+import { WithMiddleware } from "@/middlewares/WithMiddleware";
+import { CustomApiError } from "@/types/CustomApiError";
 import Room, { SchemaFields as RoomSchemaFields, RoomUpdateData } from "@/models/Room";
 import Joi from "joi";
-import { FirebaseStorage, StoragePaths } from "@/lib/firebaseAdmin/init";
-import { resizeImageOneSz } from "@/lib/utils/dataConversion";
-import { RateLimits } from "@/middlewares/rateLimit";
+import { FirebaseStorage, StoragePaths } from "@/firebase/init";
+import { resizeImageOneSz } from "@/utils/dataConversion";
+import { RateLimits } from "@/middlewares/RateLimiter";
 import { MultiSizePhoto } from "@/models/Identity";
 
 export const config = {
@@ -226,7 +226,7 @@ async function deleteImages(imagesToDelete: MultiSizePhoto[]): Promise<void> {
  * response = { message: string, imagesUpdated: boolean }
  * ```
  */
-export default withmiddleware(async function PATCH(req: NextApiRequest, res: NextApiResponse) {
+export default WithMiddleware(async function PATCH(req: NextApiRequest, res: NextApiResponse) {
   // Only allow PATCH method
   if (req.method !== "PATCH") {
     throw CustomApiError.create(405, "Method Not Allowed");
