+++ backend/src/pages/api/rooms/readListOnQuery/index.ts
@@ -1,11 +1,11 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import Room, { AcceptGender, AcceptOccupation, RoomQueryParams, RoomReadDataWithId } from "@/models/Room";
 import { Timestamp } from "firebase-admin/firestore";
-import { respond } from "@/lib/utils/respond";
-import { withmiddleware } from "@/middlewares/withMiddleware";
-import { getLoggedInUser } from "@/middlewares/auth";
-import { CustomApiError } from "@/lib/utils/ApiError";
-import { RateLimits } from "@/middlewares/rateLimit";
+import { respond } from "@/utils/respond";
+import { WithMiddleware } from "@/middlewares/WithMiddleware";
+import { getLoggedInUser } from "@/middlewares/Auth";
+import { CustomApiError } from "@/types/CustomApiError";
+import { RateLimits } from "@/middlewares/RateLimiter";
 import { LRUCache } from "lru-cache";
 
 // Configure LRU cache
@@ -64,7 +64,7 @@ const PAGE_SIZE = 8;
  * }
  * ```
  */
-export default withmiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
+export default WithMiddleware(async function GET(req: NextApiRequest, res: NextApiResponse) {
   // Allow only GET requests
   if (req.method !== "GET") {
     throw CustomApiError.create(405, "Method Not Allowed");
